Filling up the memory

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -8,9 +8,9 @@
 */
 char *_memset(char *s, char b, unsigned int n)
 {
-    int i;
+    unsigned int i;
 
-    for(i=0; i < (int) n; i++)
+    for(i=0; i <  n; i++)
     {
         s[i] = b;
     }
